requirements.txt
pyserial
lxml
cssselect
